test: update users POST e2e with basic validation and request handling

+++ test/users/users.post.e2e-spec.ts
@@ -1,3 +1,4 @@
+import  request from 'supertest';
 import { INestApplication } from '@nestjs/common';
 import { App } from 'supertest/types';
 import { ConfigService } from '@nestjs/config';
@@ -7,6 +8,7 @@ import { bootstrapNestApplication } from 'test/helpers/bootstrap-nest-applicatio
 describe('[Users] @Post Endpoints', () => {
   let app: INestApplication<App>;
   let config: ConfigService;
+  let httpServer:App
 
   beforeEach(async () => {
     // Instantiation the application
@@ -14,13 +16,21 @@ describe('[Users] @Post Endpoints', () => {
 
     // Extract the config
     config = app.get<ConfigService>(ConfigService);
+    
+    httpServer = app.getHttpServer()
   });
 
   afterEach(async () => {
-    await dropDatabase(config);
+    if(config) await dropDatabase(config);
+    if (app) await app.close();
+
   });
 
-  it.todo('/users - endpoint is public');
+  it('/users - endpoint is public', async ()=> {
+    return await request(httpServer).post('/users').send({}).expect(400).then((data)=>{
+      console.log(data.body);
+    })
+  })
   it.todo('/users - firstName is mandatory');
   it.todo('/users - email is mandatory');
   it.todo('/users - password is mandatory');
